LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/import-assertions/invalid-type-assertion-error-expected.txt


FAIL Test that invalid module type assertion leads to TypeError on window. assert_equals: expected 4 but got 8
